Consolidated everything into a single /setup page without any steps
